Rename the asset (typo)
