+++ src/pages/Error.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom'
-import Wrapper from '../asserts/wrappers/Error'
+import Wrapper from '../assets/wrappers/Error'
 
 const Error = () => {
     return (
